backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1